+++ src/test/php/filter/JsonFilterTest.php
@@ -19,6 +19,7 @@ require_once __DIR__ . '/FilterTest.php';
  * Tests for stubbles\input\filter\JsonFilter.
  *
  * @group  filter
+ * @group  json
  */
 class JsonFilterTest extends FilterTest
 {
@@ -34,7 +35,7 @@ class JsonFilterTest extends FilterTest
      */
     public function setUp()
     {
-        $this->jsonFilter = JsonFilter::instance();
+        $this->jsonFilter = new JsonFilter();
         parent::setUp();
     }
 
@@ -253,6 +254,15 @@ class JsonFilterTest extends FilterTest
     }
 
     /**
+     * @since  6.0.0
+     * @test
+     */
+    public function asJsonReturnsNullIfParamIsTooLong()
+    {
+        assertNull($this->readParam(json_encode('foo'))->asJson(1));
+    }
+
+    /**
      * @since  2.0.0
      * @test
      */
@@ -261,7 +271,6 @@ class JsonFilterTest extends FilterTest
         $this->readParam('foo')->asJson();
         assertTrue($this->paramErrors->existFor('bar'));
     }
-
     /**
      * @test
      */